refactor: CustomInputProps 인터페이스를 CustomInput.tsx로 이동
